refactor: move face to PathObject
