+++ src/renderer/cat/render_info.ts
@@ -1,0 +1,15 @@
+/**
+ * @license
+ * Copyright 2026 Scratch Foundation
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { RenderInfo as ClassicRenderInfo } from "../render_info";
+
+import { ConstantProvider } from "./constants";
+import { CatScratchRenderer } from "./renderer";
+
+export class RenderInfo extends ClassicRenderInfo {
+    override constants_: ConstantProvider;
+    override renderer_: CatScratchRenderer;
+}
